Move CsvLine to infrastructure/utils/CampaignAssessmentCsvLine
